refactor: extract shared catalog search adapter to fix duplication
Move catalogRowToSearchItem and findLatestItemsFromCatalog from CLI
and serve handlers into a shared adapter at
src/infrastructure/persistence/catalog_search_adapter.ts.

Addresses code review finding: duplicated ~50 lines across two files
was a maintenance hazard where updates to the mapping in one file
could silently diverge from the other.
